+++ src/js/components/Window.jsx
@@ -1,14 +1,12 @@
 /**
  * Created by Eric on 9/6/2018.
  */
-import React, { lazy, Suspense } from "react";
+import React from "react";
 import { tileToBBox, pointToTileFraction, googleToTile } from "global-mercator";
 import { ActionWatcher } from "./utils";
 import { TopComp, TopComp1 } from "../context/ContextShow";
 import { ValidationWrapper } from "./ValidationTemp";
 import DemoDynamicImport from "./demo-dynamic-import";
-const NamedLazyComp = lazy(() => import("./demo-fake").then(module => ({ default: module.Fake })));
-const DefaultLazyComp = lazy(() => import("./demo-fake"));
 import custom from "../../css/main.css"; // eslint-disable-line
 import bootstrap from "bootstrap/dist/css/bootstrap.min.css"; // eslint-disable-line
 import fontawesome from "font-awesome/css/font-awesome.min.css"; // eslint-disable-line
@@ -104,14 +102,7 @@ export class Window extends React.Component {
         <hr />
         <ValidationWrapper />
         <hr />
-        <div>
-          <DemoDynamicImport />
-          <h6>React Lazy Suspense Dynamic Import</h6>
-          <Suspense fallback={<div>MyLazyComp is Loading</div>}>
-            <NamedLazyComp />
-            <DefaultLazyComp />
-          </Suspense>
-        </div>
+        <DemoDynamicImport />
         <hr />
         <TopComp myMessage={"hello world"} />
         <TopComp1 />
